refactor: 플레이어 이름 값객체로 분리
